requirements.txt
langchain==0.3.26
flask==3.1.1
flask-cors
pypdf
python-dotenv
langchain-pinecone==0.2.8
langchain-core>=0.3.66
langchain-groq==0.3.2
langchain-community==0.3.26
langchain-huggingface==0.1.2
huggingface_hub>=0.23.0
pinecone
gunicorn
requests
